+++ classes/Series/class.xoctSeriesFormGUI.php
@@ -36,6 +36,8 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	const F_SHOW_UPLOAD_TOKEN = 'show_upload_token';
 	const F_PUBLISH_ON_VIDEO_PORTAL = 'publish_on_video_portal';
 	const F_PERMISSION_TEMPLATE = 'permission_template';
+	const F_DEFAULT_VIEW = 'default_view';
+	const F_VIEW_CHANGEABLE = 'view_changeable';
 
 	/**
 	 * @var  xoctSeries
@@ -65,6 +67,9 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	 * @param bool $view
 	 * @param bool $infopage
 	 * @param bool $external
+	 * @throws \srag\DIC\OpenCast\Exception\DICException
+	 * @throws arException
+	 * @throws xoctException
 	 */
 	public function __construct($parent_gui, xoctOpenCast $cast, $view = false, $infopage = false, $external = true) {
 		parent::__construct();
@@ -87,6 +92,10 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	}
 
 
+	/**
+	 * @throws \srag\DIC\OpenCast\Exception\DICException
+	 * @throws arException
+	 */
 	protected function initForm() {
 		$xoctUser = xoctUser::getInstance(self::dic()->user());
 		$this->setTarget('_top');
@@ -149,12 +158,6 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 		$te->setRows(5);
 		$this->addItem($te);
 
-		//		$discipline = new ilSelectInputGUI($this->txt(self::F_DISCIPLINE), self::F_DISCIPLINE);
-		//		sort(self::$disciplines);
-		//		$discipline->setOptions(self::$disciplines);
-		//		$discipline->setRequired(true);
-		//		$this->addItem($discipline);
-
 		$license = new ilSelectInputGUI($this->txt(self::F_LICENSE), self::F_LICENSE);
 		$options = array(
 			null => 'As defined in content',
@@ -172,6 +175,22 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 		$license->setInfo($license_info);
 		$license->setOptions($options);
 		$this->addItem($license);
+
+
+		if (xoct::ILIAS_54) {
+			$default_view = new ilSelectInputGUI($this->txt(self::F_DEFAULT_VIEW), self::F_DEFAULT_VIEW);
+			$options = [
+				xoctUserSettings::VIEW_TYPE_LIST => $this->txt('view_type_' . xoctUserSettings::VIEW_TYPE_LIST),
+				xoctUserSettings::VIEW_TYPE_TILES => $this->txt('view_type_' . xoctUserSettings::VIEW_TYPE_TILES),
+			];
+			$default_view->setOptions($options);
+			$this->addItem($default_view);
+
+
+			$view_changeable = new ilCheckboxInputGUI($this->txt(self::F_VIEW_CHANGEABLE), self::F_VIEW_CHANGEABLE);
+			$view_changeable->setInfo($this->txt(self::F_VIEW_CHANGEABLE . '_info'));
+			$this->addItem($view_changeable);
+		}
 
 		$department = new ilTextInputGUI($this->txt(self::F_DEPARTMENT), self::F_DEPARTMENT);
 		$department->setInfo($this->infoTxt(self::F_DEPARTMENT));
@@ -236,8 +255,11 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	}
 
 
+	/**
+	 *
+	 */
 	public function fillFormRandomized() {
-		$array = array(
+		$this->setValuesByArray([
 			self::F_CHANNEL_TYPE             => self::EXISTING_NO,
 			self::F_TITLE                    => 'New Channel ' . date(DATE_ATOM),
 			self::F_DESCRIPTION              => 'This is a description',
@@ -248,15 +270,15 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 			self::F_PERMISSION_PER_CLIP      => true,
 			self::F_PERMISSION_ALLOW_SET_OWN => true,
 			self::F_ACCEPT_EULA              => true,
-		);
-
-		$this->setValuesByArray($array);
+		]);
 	}
 
 
+	/**
+	 *
+	 */
 	public function fillForm() {
-
-		$array = array(
+		$array = [
 			self::F_CHANNEL_TYPE             => self::EXISTING_NO,
 			self::F_TITLE                    => $this->series->getTitle(),
 			self::F_DESCRIPTION              => $this->series->getDescription(),
@@ -269,9 +291,12 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 			self::F_OBJ_ONLINE               => $this->cast->isOnline(),
 			self::F_CHANNEL_ID               => $this->cast->getSeriesIdentifier(),
 			self::F_PERMISSION_TEMPLATE      => $this->series->getPermissionTemplateId(),
-            self::F_PUBLISH_ON_VIDEO_PORTAL  => $this->series->isPublishedOnVideoPortal(),
-		);
-
+			self::F_PUBLISH_ON_VIDEO_PORTAL  => $this->series->isPublishedOnVideoPortal(),
+			self::F_DEFAULT_VIEW  			 => $this->cast->getDefaultView(),
+		];
+		if (xoct::isIlias54()) {
+			$array[self::F_VIEW_CHANGEABLE] = $this->cast->isViewChangeable();
+		}
 		$this->setValuesByArray($array);
 	}
 
@@ -280,6 +305,7 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	 * returns whether checkinput was successful or not.
 	 *
 	 * @return bool
+	 * @throws xoctException
 	 */
 	public function fillObject() {
 		if (!$this->checkInput()) {
@@ -306,6 +332,10 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 		$this->cast->setPermissionAllowSetOwn($this->getInput(self::F_PERMISSION_ALLOW_SET_OWN));
 		$this->cast->setOnline($this->getInput(self::F_OBJ_ONLINE));
 		$this->cast->setAgreementAccepted(true);
+		if (xoct::isIlias54()) {
+			$this->cast->setDefaultView($this->getInput(self::F_DEFAULT_VIEW));
+			$this->cast->setViewChangeable($this->getInput(self::F_VIEW_CHANGEABLE));
+		}
 
 		return true;
 	}
@@ -315,6 +345,7 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	 * @param $key
 	 *
 	 * @return string
+	 * @throws \srag\DIC\OpenCast\Exception\DICException
 	 */
 	protected function txt($key) {
 		return self::plugin()->getPluginObject()->txt('series_' . $key);
@@ -325,6 +356,7 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	 * @param $key
 	 *
 	 * @return string
+	 * @throws \srag\DIC\OpenCast\Exception\DICException
 	 */
 	protected function infoTxt($key) {
 		return self::plugin()->getPluginObject()->txt('series_' . $key . '_info');
@@ -395,6 +427,9 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	}
 
 
+	/**
+	 *
+	 */
 	protected function initButtons() {
 		if ($this->is_new) {
 			$this->setTitle($this->txt('create'));
@@ -421,6 +456,10 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 	}
 
 
+	/**
+	 * @throws \srag\DIC\OpenCast\Exception\DICException
+	 * @throws arException
+	 */
 	protected function initView() {
 		$this->initForm();
 		/**
@@ -455,115 +494,9 @@ class xoctSeriesFormGUI extends ilPropertyFormGUI {
 		$this->series = $series;
 	}
 
-
-	protected static $disciplines = array(
-		1932 => 'Arts & Culture',
-		5314 => 'Architecture',
-		6302 => 'Landscape architecture',
-		5575 => 'Spatial planning',
-		9202 => 'Art history',
-		3119 => 'Design',
-		6095 => 'Industrial design',
-		5103 => 'Visual communication',
-		5395 => 'Film',
-		8202 => 'Music',
-		2043 => 'Music education',
-		9610 => 'School and church music',
-		3829 => 'Theatre',
-		1497 => 'Visual arts',
-		6950 => 'Business',
-		1676 => 'Business Administration',
-		4949 => 'Business Informatics',
-		7290 => 'Economics',
-		2108 => 'Facility Management',
-		7641 => 'Hotel business',
-		6238 => 'Tourism',
-		5214 => 'Education',
-		1672 => 'Logopedics',
-		1406 => 'Pedagogy',
-		3822 => 'Orthopedagogy',
-		2150 => 'Special education',
-		9955 => 'Teacher education',
-		6409 => 'Primary school',
-		7008 => 'Secondary school I',
-		4233 => 'Secondary school II',
-		8220 => 'Health',
-		2075 => 'Dentistry',
-		5955 => 'Human medicine',
-		5516 => 'Nursing',
-		3424 => 'Pharmacy',
-		4864 => 'Therapy',
-		6688 => 'Occupational therapy',
-		7072 => 'Physiotherapy',
-		3787 => 'Veterinary medicine',
-		4832 => 'Humanities',
-		1438 => 'Archeology',
-		8796 => 'History',
-		7210 => 'Linguistics & Literature (LL)',
-		9557 => 'Classical European languages',
-		9391 => 'English LL',
-		9472 => 'French LL',
-		4391 => 'German LL',
-		3468 => 'Italian LL',
-		7408 => 'Linguistics',
-		6230 => 'Other modern European languages',
-		5676 => 'Other non-European languages',
-		5424 => 'Rhaeto-Romanic LL',
-		7599 => 'Translation studies',
-		7258 => 'Musicology',
-		4761 => 'Philosophy',
-		3867 => 'Theology',
-		6527 => 'General theology',
-		5633 => 'Protestant theology',
-		9787 => 'Roman catholic theology',
-		5889 => 'Interdisciplinary & Other',
-		6059 => 'Information & documentation',
-		5561 => 'Military sciences',
-		8683 => 'Sport',
-		1861 => 'Law',
-		4890 => 'Business law',
-		2990 => 'Natural sciences & Mathematics',
-		8990 => 'Astronomy',
-		4195 => 'Biology',
-		7793 => 'Ecology',
-		6451 => 'Chemistry',
-		1266 => 'Computer science',
-		5255 => 'Earth Sciences',
-		7950 => 'Geography',
-		2158 => 'Mathematics',
-		6986 => 'Physics',
-		8637 => 'Social sciences',
-		9619 => 'Communication and media studies',
-		8367 => 'Ethnology',
-		1774 => 'Gender studies',
-		1514 => 'Political science',
-		6005 => 'Psychology',
-		7288 => 'Social work',
-		6525 => 'Sociology',
-		9321 => 'Technology & Applied sciences',
-		3624 => 'Agriculture',
-		1442 => 'Enology',
-		1892 => 'Biotechnology',
-		7132 => 'Building Engineering',
-		5727 => 'Chemical Engineering',
-		9389 => 'Construction Science',
-		2527 => 'Civil Engineering',
-		9738 => 'Rural Engineering and Surveying',
-		5742 => 'Electrical Engineering',
-		2850 => 'Environmental Engineering',
-		9768 => 'Food technology',
-		2979 => 'Forestry',
-		1566 => 'Material sciences',
-		8189 => 'Mechanical Engineering',
-		5324 => 'Automoive Engineering',
-		8502 => 'Microtechnology',
-		4380 => 'Production and Enterprise',
-		7303 => 'Telecommunication',
-	);
-
-
 	/**
 	 * @return bool
+	 * @throws \srag\DIC\OpenCast\Exception\DICException
 	 */
 	protected function checkEula() {
 		if ($this->is_new && !$this->getInput(self::F_ACCEPT_EULA)) {