+++ plugin.php
@@ -1,7 +1,7 @@
 <?php
 $id = 'xoct';
-$version = '2.2.1';
-$ilias_min_version = '5.2.0';
-$ilias_max_version = '5.3.999';
+$version = '2.3.0';
+$ilias_min_version = '5.3.0';
+$ilias_max_version = '5.4.999';
 $responsible = 'Theodor Truffer';
 $responsible_mail = 'support@studer-raimann.ch';