+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [2.3.0]
+- Support for ILIAS 5.4
+- Dropped support for ILIAS 5.2
+- Feature: Tile View
+- Feature: Set field 'presenter/s' required for scheduled events
+
 ## [2.2.1]
 - Bugfix: Config import not working